Rename generated package path
