+++ api/applyconfigurations/gardener/cert/v1alpha1/certificatespec.go
