+++ api/applyconfigurations/gardener/cert/v1alpha1/pkcs12keystore.go
