+++ api/applyconfigurations/gardener/cert/v1alpha1/qualifiedissuerref.go
